+++ test/strategy.test.js
@@ -26,13 +26,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: undefined,
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -62,13 +56,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -101,13 +89,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -138,13 +120,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -175,13 +151,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -212,13 +182,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -249,13 +213,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -286,13 +244,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -323,13 +275,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -360,13 +306,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -397,13 +337,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -434,13 +368,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -471,13 +399,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -508,13 +430,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })
@@ -550,13 +466,7 @@ describe('Strategy', function() {
         expect(state).to.have.length(24);
         expect(this.session['openidconnect:server.example.com'].state).to.deep.equal({
           handle: state,
-          issuer: 'https://server.example.com',
-          authorizationURL: 'https://server.example.com/authorize',
-          tokenURL: 'https://server.example.com/token',
-          userInfoURL: undefined,
-          clientID: 's6BhdRkqt3',
-          callbackURL: 'https://client.example.org/cb',
-          customHeaders: undefined
+          issuer: 'https://server.example.com'
         });
         done();
       })